% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_streamcat_data.R
\name{get_streamcat_data}
\alias{get_streamcat_data}
\title{Get StreamCat data}
\usage{
get_streamcat_data(metric = NA, aoi = NA, comid = NA, state = NA,
  county = NA, region = NA, showPctFull = NA)
}
\arguments{
\item{metric}{Names of a metric to query
Syntax: name=<name1>,<name2>}

\item{aoi}{Specify the area of interest described by a metric. By default, all available areas of interest 
for a given metric are returned.
Syntax: areaOfInterest=<value1>,<value2>
Values: catchment|watershed|riparian_catchment|riparian_watershed|other}

\item{comid}{Return metric information about specific COMIDs
Syntax: comid=<comid1>,<comid2>}

\item{state}{Return metric information for COMIDs within a specific state. Use a state's abbreviation to 
query for a given state.
Syntax: state=<state1>,<state2>}

\item{county}{Return metric information for COMIDs within a specific county. 
Users must use FIPS code, not county name, as a way to disambiguate counties.
Syntax: county=<county1>,<county1>}

\item{region}{Return metric information for COMIDs within a specified hydroregion.
Syntax: region=<regionid1>,<regionid2>}

\item{showPctFull}{Return the "*pctfull" row for each dataset. The default value is false.
Values: true|false}
}
\value{
A tibble of desired StreamCat metrics
}
\description{
Function to return StreamCat catchment and watershed metrics using the StreamCat API.  The function allows a user to get data
specific to a catchment or metric of interest, as well as returning data by hydroregion(s), state(s), or county(ies).
}
\examples{
df <- get_streamcat_data(comid='179', aoi='catchment', metric='fert')

df <- get_streamcat_data(metric='PctGrs2006', aoi='watershed', region='01')

df <- get_streamcat_data(metric='PctUrbMd2006', aoi='riparian_catchment', comid='1337420')

df <- get_streamcat_data(metric='PctUrbMd2006,DamDens,TRIDens', aoi='riparian_catchment,catchment,watershed', comid='179,1337,1337420')
}
