% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_streamcat_params.R
\name{get_streamcat_params}
\alias{get_streamcat_params}
\title{Get StreamCat Parameters}
\usage{
get_streamcat_params(param = NULL)
}
\arguments{
\item{param}{Either name or area to grab JSON of parameters in API
Syntax: param=<value1>,<value2>
Values: name|area}
}
\value{
A list of all the current StreamCat values for a given parameter
}
\description{
Function to return available StreamCat parameters using the StreamCat API.
}
\examples{
params <- get_streamcat_params(param='name')
params <- get_streamcat_params(param='area')
}
\author{
Marc Weber
}
