% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_get_params.R
\name{sc_fullname}
\alias{sc_fullname}
\title{Lookup Full Metric Name}
\usage{
sc_fullname(metric = NULL)
}
\arguments{
\item{metric}{Short metric name
Syntax: metric=value1
Values: metric}
}
\value{
A lookup of the full name for a given StreamCat metric
}
\description{
Function to retrieve a full metric name based on the short name using the StreamCat API.
}
\examples{
\donttest{
fullname <- sc_fullname(metric='clay')
}
}
\author{
Marc Weber
}
