% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_get_data.R
\name{sc_get_nni}
\alias{sc_get_nni}
\title{Get NNI}
\usage{
sc_get_nni(
  year,
  aoi = NULL,
  comid = NULL,
  showAreaSqKm = TRUE,
  state = NULL,
  county = NULL,
  region = NULL,
  conus = NULL,
  showPctFull = NULL,
  countOnly = NULL
)
}
\arguments{
\item{year}{Years(s) of NNI metrics to query.
Only valid NNI years are accepted (1987:2017)
Syntax: year=<year1>,<year2>}

\item{aoi}{Specify the area of interest described by a metric. By default, all available areas of interest
for a given metric are returned.
Syntax: areaOfInterest=<value1>,<value2>
Values: catchment|watershed}

\item{comid}{Return metric information for specific COMIDs
Syntax: comid=<comid1>,<comid2>}

\item{showAreaSqKm}{Return the area in square kilometers of a given area of interest.
The default value is true.
Values: true|false}

\item{showPctFull}{Return the pctfull for each dataset. The default value is false.
Values: true|false}

\item{countOnly}{Return a CSV containing only the row count (ROWCOUNT) and the column
count (COLUMNCOUNT) that the server expects to return in a request. The default value is false.
Values: true|false}
}
\value{
A tibble of desired StreamCat metrics
}
\description{
Function to get all NNI data available for a given year.
}
\author{
Selia Markley
}
