% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_get_params.R
\name{lc_get_params}
\alias{lc_get_params}
\title{Get LakeCat Parameters}
\usage{
lc_get_params(param = NULL)
}
\arguments{
\item{param}{List of available parameters in the API for the following options:
name, areaofInterest, region, state, county.  State and county return a data
frame that includes FIPS codes, names and state abbreviations
Syntax: param=<value1>,<value2>
Values: name|area}
}
\value{
A list of all the current LakeCat values for a given parameter
}
\description{
Function to return available LakeCat parameters using the StreamCat API.
}
\examples{
\dontrun{
params <- lc_get_params(param='variable_info')
params <- lc_get_params(param='metric_names')
params <- sc_get_params(param='categories')
params <- lc_get_params(param='aoi')
params <- lc_get_params(param='state')
params <- lc_get_params(param='county')
params <- sc_get_params(param='datasets')
}
}
\author{
Marc Weber
}
