% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_get_comid.R
\name{sc_get_comid}
\alias{sc_get_comid}
\title{Get COMIDs}
\usage{
sc_get_comid(dd = NULL, xcoord = NULL, ycoord = NULL, crsys = NULL)
}
\arguments{
\item{dd}{Name of data frame object.  Can be a simple
data frame with coordinate columns in a known CRS or
an sf points data frame}

\item{xcoord}{The x coordinate column if using a raw data
frame}

\item{ycoord}{The y coordinate column if using a raw data 
frame}

\item{crsys}{The epsg code if using a raw data frame}
}
\value{
A new sf data frame with a populated 'COMID' column
}
\description{
Function to return NHDPlusV2 COMIDS using either a 
dataframe with coordinates and a specified CRS or an 
sf object.  The function generates a vector of COMID
values a user can then pass to sc_get_data function
}
\examples{
\donttest{
set.seed(1234)
pt <- 10000
dd <- data.frame(x = runif(pt, 0, 100),
                 y = runif(pt, 0,50),
                 f1 = rnorm(pt))
                                  
comids <- sc_get_comid(dd, xcoord='x', 
ycoord='y', crsys=4269)

dd <- sf::read_sf(system.file("shape/nc.shp", package="sf"))
comids <- sc_get_comid(dd)

 }
}
\author{
Marc Weber
}
