% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_get_comid.R
\name{sc_get_comid}
\alias{sc_get_comid}
\title{Get COMIDs}
\usage{
sc_get_comid(df = NULL, xcoord = NULL, ycoord = NULL, crdsys = NULL)
}
\arguments{
\item{df}{Name of data frame object.  Can be a simple
data frame with coordinate columns in a known CRS or
an sf points data frame}

\item{xcoord}{The x coordinate column if using a raw data
frame}

\item{ycoord}{The y coordinate column if using a raw data 
frame}

\item{crdsys}{The epsg code if using a raw data frame}
}
\value{
A string of COMIDS to pass to sc_get_data
}
\description{
Function to return NHDPlusV2 COMIDS using either a 
dataframe with coordinates and a specified CRS or an 
sf object.  The function generates a vector of COMID
values a user can then pass to sc_get_data function
}
\examples{
comids <- sc_get_comid(df=df, xcoord='lon_dd', 
ycoord='lat_dd', crdsys=4269)

comids <- sc_get_comid(df=df)

}
\author{
Marc Weber
}
